@@ -6,6 +6,7 @@
     "dev": "bun --bun vite --host",
     "build": "bun vite build",
     "preview": "vite preview",
+    "lint": "eslint -c .eslintrc.js --ext .js,.jsx,.ts,.tsx,.vue ./src --fix",
     "ionic:build": "npm run build",
     "ionic:serve": "npm run dev"
   },
@@ -24,24 +25,24 @@
     "vuetify": "^3.3.22"
   },
   "devDependencies": {
+    "@feature-sliced/eslint-config": "^0.1.0-beta.6",
     "@mdi/font": "^7.3.67",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
     "@vitejs/plugin-vue": "^4.2.3",
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-plugin-typescript": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
     "@vue/eslint-config-airbnb": "^6.0.0",
-    "@vue/eslint-config-typescript": "^9.1.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
     "babel-eslint": "^10.1.0",
     "bun-types": "latest",
-    "eslint": "^8.50.0",
+    "eslint": "^8.52.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-vue": "^9.18.0",
     "eslint-plugin-vuejs-accessibility": "^1.1.0",
     "prettier": "^3.0.3",
     "sass": "^1.32.7",
