@@ -1,9 +1,9 @@
 import '@mdi/font/css/materialdesignicons.css'; // Ensure you are using css-loader
 import 'vuetify/styles';
-import { VDatePicker } from 'vuetify/labs/VDatePicker';
 import { createVuetify } from 'vuetify';
 import * as components from 'vuetify/components';
 import * as directives from 'vuetify/directives';
+import { VDatePicker } from 'vuetify/labs/VDatePicker';
 
 export const vuetify = createVuetify({
   components: {
