@@ -1,8 +1,6 @@
-<script setup lang="ts"></script>
-
 <template>
   <v-date-picker>
-    <v-spacer></v-spacer>
+    <!-- <v-spacer></v-spacer> -->
 
     <v-btn>Open</v-btn>
   </v-date-picker>
